+++ src/conversions/mod.rs
@@ -1,5 +1,6 @@
 //! This module contains conversions between various Rust object and their representation in Python.
 
+pub mod anyhow;
 mod array;
 pub mod eyre;
 pub mod hashbrown;