+++ tests/test_arithmetics.rs
@@ -366,6 +366,14 @@ impl LhsAndRhs {
     fn __rmatmul__(&self, other: &PyAny) -> String {
         format!("{:?} @ RA", other)
     }
+
+    fn __rtruediv__(&self, other: &PyAny) -> String {
+        format!("{:?} / RA", other)
+    }
+
+    fn __rfloordiv__(&self, other: &PyAny) -> String {
+        format!("{:?} // RA", other)
+    }
 }
 
 #[test]
@@ -639,6 +647,8 @@ mod return_not_implemented {
         _test_binary_dunder("rsub");
         _test_binary_dunder("rmul");
         _test_binary_dunder("rmatmul");
+        _test_binary_dunder("rtruediv");
+        _test_binary_dunder("rfloordiv");
         _test_binary_dunder("rmod");
         _test_binary_dunder("rdivmod");
         _test_binary_dunder("rpow");